cash: add `reset_cash_unit_counts` function
Adds a function to reset the logical and physical cash unit counts to
zero. This is useful after physically removing cash from the unit, since
the device does not automatically reset the counts.

Extends the interface for `LogicalCashUnit` and `PhysicalCashUnit` with
accessor functions for all of the fields in the structs.