fixup/currency: fix incorrect values, add accessors
Fixes an incorrect field in `DenominationItem`. `DenominationItem::unit`
should be a `u32` as it represents the unique number identifying a
`LogicalCashUnit`, not an `LCU` variant.

Adds accessor functions to `DenominationItem` for external library users
to access and set internal fields.

Adds convenience functions to `Currency` for converting to/from MDU
values.