fixup: improve XML parsing to handle XFS
Fixes the data layout of Rust XFS types for better compatibility with
`serde-xml-rs`. Items included in collections need to be wrapped in
`enum` types.

Collections should use the attribute:

`#[serde(rename = "$value", default)]`

This indicates fields should be renamed to the wrapped type's
serialized name, and empty lists are allowed.