bnr-xfs: replace `bnr-sys` with `bnr-xfs`
Replaces the C wrapper library `bnr-sys` with `bnr-xfs`.

Removes the majority of C code from the project. There is still an
indirect dependency on C via the `rusb` dependency wrapper around
`libusb`.

This is still a huge improvement on the amount of memory-safe code,
since the MEI C libraries also wrap `libusb`.